Add storage-class memory-semantics bit to LLVM IR atomic translation
